--- a/python/sglang/srt/layers/attention/flashattention_backend.py
+++ b/python/sglang/srt/layers/attention/flashattention_backend.py
@@ -97,8 +97,6 @@ class FlashAttentionBackend(AttentionBackend):
                     ),
                     (1, 0),
                 )
-                # .repeat_interleave(self.topk) # tensor([7, 7, 7, 8, 8, 8])
-                # .repeat(self.topk) # tensor([7, 8, 7, 8, 7, 8])
                 metadata.max_seq_len_k = forward_batch.seq_lens_cpu.max().item() + (
                     self.step_id + 1
                 )
@@ -112,15 +110,22 @@ class FlashAttentionBackend(AttentionBackend):
                     self.speculative_num_steps, -1
                 ).T
 
-                for idx, single_seq_len in enumerate(seq_lens_with_decode):
-                    real_bsz_start_idx = idx * self.topk
-                    real_bsz_end_idx = (idx + 1) * self.topk
-                    metadata.page_table[
-                        real_bsz_start_idx:real_bsz_end_idx,
-                        (single_seq_len - (self.step_id + 1)) : single_seq_len,
-                    ] = cache_loc[
-                        real_bsz_start_idx:real_bsz_end_idx, : (self.step_id + 1)
-                    ]
+                # page table indices to update
+                # [bsz, topk]
+                row_indices = torch.arange(batch_size * self.topk, device=device, dtype=torch.int32).view(batch_size, self.topk)
+                # [max_seq_len : max_seq_len + step_id + 1]
+                col_indices = torch.arange(forward_batch.seq_lens_cpu.max().item(), metadata.max_seq_len_k, device=device, dtype=torch.int32)
+                # mask for all valid page table indices
+                valid_mask = (col_indices.view(1, -1) >= seqlens_in_batch.view(-1, 1)) & (col_indices.view(1, -1) < seq_lens_with_decode.view(-1, 1))
+
+                # cache indices to read
+                cache_indices = torch.arange(self.step_id + 1, device=device, dtype=torch.int32)
+
+                metadata.page_table[row_indices, col_indices] = torch.where(
+                    valid_mask,
+                    cache_loc[row_indices, cache_indices].to(torch.int32),
+                    metadata.page_table[row_indices, col_indices]
+                )
             else:
                 metadata.cu_seqlens_q = torch.arange(
                     0, batch_size + 1, dtype=torch.int32, device=device