… python/sglang/srt/speculative/eagle_worker.py