<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation

Support Spec Decoding for FA3 Backend


## Benchmark on Latency


**Baseline**

```bash
python3 -m sglang.launch_server --model /shared/public/elr-models/m
eta-llama/Meta-Llama-3.1-8B-Instruct/07eb05b21d191a58c577b4a45982fe0c049d0693  --dtype float16 -
-trust-remote-code --attention-backend fa3 --disable-cuda-graph

python3 /home/jobuser/sglang/python/sglang/test/send_one.py

acc_length=1.00
speed=84.82 token/s
```

**With Spec Decoding**
```bash
python3 -m sglang.launch_server --model /shared/public/elr-models/meta-llama/Meta-Llama-3.1-8B-Instruct/07eb05b21d191a58c577b4a45982fe0c049d0693  --speculative-algorithm EAGLE3     --speculative-draft-model-path /shared/public/elr-models/jamesliu1/sglang-EAGLE3-Llama-3.1-Instruct-8B/e5ed08d66f528a95ce89f5d4fd136a28f6def714 --speculative-num-steps 4 --speculative-eagle-topk 2 --speculative-num-draft-tokens 4 --mem-fraction 0.6         --cuda-graph-max-bs 2 --dtype float16 --trust-remote-code --attention-backend fa3 --disable-cuda-graph

python3 /home/jobuser/sglang/python/sglang/test/send_one.py

acc_length=2.20
speed=125.59 token/s
```

**Sanity Check about FA3 without decoding's latency with Cuda Graph**
```bash
python -m sglang.bench_one_batch --model /shared/public/elr-models/meta-llama/Meta-Llama-3.1-8B-Instruct/07eb05b21d191a58c577b4a45982fe0c049d0693  --batch-size 16 --input 1024 --output 512 --attention-backend fa3

Benchmark ...
Prefill. latency: 0.36639 s, throughput:  44717.26 token/s
Decode.  median latency: 0.00657 s, median throughput:   2434.57 token/s
Total. latency:  3.729 s, throughput:   6589.93 token/s
```

**Sanity Check about FA3 without decoding's latency without Cuda Graph**
```bash
python -m sglang.bench_one_batch --model /shared/public/elr-models/meta-llama/Meta-Llama-3.1-8B-Instruct/07eb05b21d191a58c577b4a45982fe0c049d0693  --batch-size 16 --input 1024 --output 512 --attention-backend fa3

Prefill. latency: 0.36539 s, throughput:  44839.46 token/s
Decode.  median latency: 0.01224 s, median throughput:   1307.66 token/s
Total. latency:  6.661 s, throughput:   3689.27 token/s
```


## Benchmark on Accuracy (GSM8K)
Config: `--speculative-num-steps 4 --speculative-eagle-topk 2 --speculative-num-draft-tokens 4`
```bash

Accuracy: 0.775
Invalid: 0.000
Latency: 12.063 s
Output throughput: 1395.627 token/s

Accuracy: 0.775
Invalid: 0.000
Latency: 8.072 s
Output throughput: 2091.660 token/s

Accuracy: 0.775
Invalid: 0.000
Latency: 8.065 s
Output throughput: 2089.109 token/s
```


**Todo in next PR:**

- Support Cuda Graph for Spec Decoding
- Support Page Size > 1 for Spec Decoding

> Currently if user tried to spawn sglang fa3 without `--disable-cuda-graph`, it will error out with:
> `NotImplementedError: FlashAttentionBackend does not support CUDA graph yet, stay tuned!`
> 




<!-- Explain the purpose of this PR and the goals it aims to achieve. -->
Add Eagle Speculative Decoding to FA3 Backend.

co-author with @zcnrex @Fridge003 @hebiao064 

## Modifications

<!-- Describe the changes made in this PR. -->

## Checklist

- [ ] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [ ] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.