@hebiao064 2025-04-02
2 tests failed due to the accuracy slightly below threshold
- https://github.com/sgl-project/sglang/actions/runs/14212153877/job/39821214745?pr=4951
- https://github.com/sgl-project/sglang/actions/runs/14212153913/job/39821214785?pr=4951
another 3 tests failed due to timeout (> 30mins)
- https://github.com/sgl-project/sglang/actions/runs/14212153877/job/39821214752?pr=4951
- https://github.com/sgl-project/sglang/actions/runs/14212153877/job/39821214749?pr=4951
- https://github.com/sgl-project/sglang/actions/runs/14212153877/job/39821214769?pr=4951